requirements.txt
streamlit
sentence-transformers
endee
python-dotenv
langchain-groq
langchain-core